Implement route handlers for login and logout processes
