+++ app/api/logout/route.js
@@ -1,0 +1,9 @@
+import { cookies } from "next/headers";
+import { AUTH_COOKIE_KEY } from "../../../constants";
+
+export async function POST() {
+  const cookieStore = cookies();
+
+  cookieStore.delete(AUTH_COOKIE_KEY);
+  return Response.json({ message: "success" });
+}
