+++ components/UI/LogoutButton.jsx
@@ -1,8 +1,17 @@
 "use client";
 
-export default function LogoutButton({ handleLogout }) {
+import { handleLogout } from "../../scripts/logout";
+
+export default function LogoutButton() {
   return (
-    <button onClick={() => handleLogout()} className="button logout-button">
+    <button
+      onClick={() =>
+        handleLogout().then(() => {
+          window.location.reload();
+        })
+      }
+      className="button logout-button"
+    >
       Logout
     </button>
   );
