+++ components/header/Header.jsx
@@ -1,15 +1,9 @@
 import Link from "next/link";
 import ToggleThemeButton from "../toggleTheme/ToggleTheme";
 import LogoutButton from "../UI/LogoutButton";
-import { logout } from "../../app/actions";
 import "./Header.css";
 
 export default function Header() {
-  const handleLogout = async () => {
-    "use server";
-    await logout();
-  };
-
   return (
     <header className="header">
       <ToggleThemeButton />
@@ -43,7 +37,7 @@ export default function Header() {
             <circle cx="12" cy="7" r="4"></circle>
           </svg>
         </Link>
-        <LogoutButton handleLogout={handleLogout} />
+        <LogoutButton />
       </nav>
     </header>
   );
