+++ constants.js
@@ -1,1 +1,2 @@
 export const AUTH_COOKIE_KEY = "Auth";
+export const TOKEN_COOKIE_KEY = 'auth_token';
