+++ scripts/login.js
@@ -1,0 +1,10 @@
+export const handleLogin = async (username, password) => {
+    const response = await fetch("http://localhost:3000/api/login", {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify({
+        username,
+        password,
+      }),
+    });
+  };
