+++ scripts/logout.js
@@ -1,0 +1,5 @@
+export const handleLogout = async () => {
+  const response = await fetch("http://localhost:3000/api/logout", {
+    method: "POST",
+  });
+};
